Add new componebt
