+++ client-app/src/features/activities/dashboard/ActivityList.tsx
@@ -1,44 +1,20 @@
 
-import { SyntheticEvent, useState } from "react";
-import { Button, Item, Label, Segment } from "semantic-ui-react";
+import { Item, Segment } from "semantic-ui-react";
 import { useStore } from "../../../app/stores/store";
 import { observer } from "mobx-react-lite";
-import { Link } from "react-router-dom";
+import ActivityListItem from "./ActivityListItems";
 
 export default observer(function ActivityList() {
-    const [target, setTarget] = useState('');
     const { activityStore } = useStore();
-    const { deleteActivity, activitiesByDate, loading } = activityStore;
+    const {activitiesByDate} = activityStore;
 
-    function handleActivityDelete(e: SyntheticEvent<HTMLButtonElement>, id: string) {
-        setTarget(e.currentTarget.name);
-        deleteActivity(id);
-    }
+
 
     return (
         <Segment>
             <Item.Group divided>
                 {activitiesByDate.map(activity => (
-                    <Item key={activity.id}>
-                        <Item.Content>
-                            <Item.Header as='a'>{activity.title}</Item.Header>
-                            <Item.Meta >{activity.date}</Item.Meta>
-                            <Item.Description>
-                                <div>{activity.description}</div>
-                                <div>{activity.city}, {activity.venue}</div>
-                            </Item.Description>
-                            <Item.Extra>
-                                <Button as={Link} to={`/activities/${activity.id}`}
-                                    floated="right" content='View' color='blue' ></Button>
-                                <Button
-                                    name={activity.id}
-                                    loading={loading && target === activity.id}
-                                    onClick={(e) => handleActivityDelete(e, activity.id)}
-                                    floated="right" content='Delete' color='red' ></Button>
-                                <Label basic content={activity.category} ></Label>
-                            </Item.Extra>
-                        </Item.Content>
-                    </Item>
+                    <ActivityListItem key={activity.id} activity={activity}/>
                 ))}
             </Item.Group>
         </Segment>
